+++ NEWS.md
@@ -1,6 +1,6 @@
 # LikertMakeR (development version)
 
-# LikertMakeR 1.4.0 (December 2025)
+# LikertMakeR 1.4.0 (February 2026)
 
 ## Improvements
 
@@ -20,6 +20,11 @@
   - coefficient H
   
   with bootstrapped Confidence Intervals for each as an option.
+
+- correlateScales() function runs 3 to 8 times faster
+
+  - due to some pre-calculation and vectorisation replacing "for loops"
+
   
 ## Maintenance